Move orchestrator status into right rail (#121)
