+++ docs/requirements_traceability.md
@@ -47,3 +47,4 @@ This matrix links requirement issues to implementation, tests, and validation ev
 | [#94](https://github.com/lumensparkxy/agentic_test_case_generator/issues/94) | Add an offline orchestrator validation suite and lifecycle E2E proving the orchestrator improves resumability, impact precision, governance, and auditability versus full regeneration for realistic QA lifecycle scenarios. | `scripts/evaluate_orchestrator.py`, `scripts/benchmark_orchestrator_inputs/`, `scripts/benchmark_orchestrator_expectations/`, `.github/workflows/ci.yml`, `frontend/e2e/orchestrator-lifecycle.spec.js`, codebase docs | `backend/tests/test_orchestrator_evaluation.py` covers v1 generation routing, v2 two-requirement impact precision, unchanged-test preservation, full-regenerate false updates avoided, governance blockers, and resumability; `frontend/e2e/orchestrator-lifecycle.spec.js` covers create project, generate v1, reload, change requirements, impact update, execute, review, and report. | Focused orchestrator evaluation tests passed 3 tests; strict orchestrator benchmark passed 3 fixtures with 1.00 average impact precision, 1.00 unchanged preservation, 16 full-regenerate false updates avoided, and 3/3 resumability matches; backend suite passed 281 tests; strict requirement evaluation passed at 100/85; strict generation evaluation passed at 100/90; backend Ruff lint and format checks passed; OpenAPI export wrote a 122,867-byte schema; generated frontend API contract check passed; frontend `npm ci`, lint, format check, build, focused export approval E2E, and focused orchestrator lifecycle E2E passed; execution-runtime `npm ci` and Playwright list check passed with 0 tests; `git diff --check` passed. |
 | [#117](https://github.com/lumensparkxy/agentic_test_case_generator/issues/117) | Redesign the frontend workflow shell into the Command Deck direction with no left navigation, a compact command bar, a horizontal workflow stepper, fewer visible statuses, and unchanged workflow behavior. | `frontend/src/App.jsx`, `frontend/src/components/layout/AppHeader.jsx`, `frontend/src/components/layout/WorkflowStepper.jsx`, `frontend/src/components/projects/NextActionPanel.jsx`, `frontend/src/components/projects/ProjectSummaryPanel.jsx`, `frontend/src/components/projects/RunDetailsDrawer.jsx`, `frontend/src/components/projects/OrchestratorCockpitPanel.jsx`, `frontend/src/components/projects/ProjectWorkspacePanel.jsx`, `frontend/src/components/automation/AutomationPanel.jsx`, `frontend/src/components/export/ExportPanel.jsx`, frontend layout/workflow styles, `docs/command-deck-design-qa.md` | Existing behavior-preserving frontend E2E coverage for workflow navigation, orchestrator cockpit/lifecycle, automation execution, JIRA workflow, export approval/download, report evidence, and export surfaces. | Baseline frontend build and orchestrator cockpit E2E passed before editing; final frontend build, format check, `git diff --check`, orchestrator cockpit E2E, orchestrator lifecycle E2E, multi-environment execution E2E, JIRA workflow E2E, export approval/download E2E, and report evidence E2E passed; broad live `workflow.spec.js` anonymous sign-in passed and the authenticated path reached export, then exceeded its 10-minute test budget waiting for the final download after real-service generation; product-design visual QA recorded `final result: passed` with the source Command Deck mock and local screenshot comparison. |
 | [#120](https://github.com/lumensparkxy/agentic_test_case_generator/issues/120) | Replace the horizontal workflow stepper with a Material 3-inspired left workflow navigation for Upload, Context, Template, Generate, Automation, and Export while preserving existing tab behavior. | `frontend/src/components/layout/WorkflowNavigationDrawer.jsx`, `frontend/src/App.jsx`, `frontend/src/styles/layout.css`, `frontend/src/styles/responsive.css`, navigation selector updates in `frontend/e2e/impact-update-flow.spec.js`, `frontend/e2e/multi-environment-execution.spec.js`, and `frontend/e2e/orchestrator-lifecycle.spec.js`, codebase docs | `frontend/e2e/workflow-navigation.spec.js` covers the left navigation destinations, active state, and content switching; existing E2E coverage verifies impact update, multi-environment execution, orchestrator cockpit, orchestrator lifecycle, report evidence, export approval, JIRA workflow, mobile/desktop CSS smoke, and full authenticated workflow behavior. | Focused navigation E2E passed 1 test; patched impact update, multi-environment execution, and orchestrator lifecycle E2E specs passed individually; full frontend E2E `E2E_BASE_URL=http://127.0.0.1:5173 npm run test:e2e` passed 15 tests with 39 generated cases and complete quality checks; frontend build, lint, format check, and `git diff --check` passed; backend suite passed 283 tests; strict requirement and generation evaluations passed at 100; OpenAPI export and generated frontend API contract check passed; execution-runtime Playwright list check passed with 0 tests. |
+| [#121](https://github.com/lumensparkxy/agentic_test_case_generator/issues/121) | Move project and orchestrator status, blockers, agent timeline, project evidence, and last-run details into a persistent right-side information rail while keeping cockpit actions unchanged. | `frontend/src/components/projects/ProjectInformationRail.jsx`, `frontend/src/components/projects/ProjectWorkspacePanel.jsx`, `frontend/src/components/projects/OrchestratorCockpitPanel.jsx`, `frontend/src/App.jsx`, `frontend/src/styles/layout.css`, `frontend/src/styles/workflow.css`, `frontend/src/styles/responsive.css`, updated cockpit/lifecycle/impact/execution/report E2E selectors, codebase docs | `frontend/e2e/orchestrator-cockpit.spec.js` covers action-region compatibility plus rail status, blockers, timeline, and checkpoint evidence; `frontend/e2e/report-evidence.spec.js` covers latest report evidence in the rail; `frontend/e2e/orchestrator-lifecycle.spec.js` covers refresh/review/report flow with rail state after reload; CSS smoke and full workflow E2E cover responsive layout and unchanged tab behavior. | Focused orchestrator cockpit E2E passed 2 tests; focused report evidence, orchestrator lifecycle, impact update, multi-environment execution, workflow navigation, and CSS smoke E2E passed; full frontend E2E `E2E_BASE_URL=http://127.0.0.1:5173 npm run test:e2e` passed 15 tests with 39 generated cases and complete quality checks; frontend build, lint, format check, and `git diff --check` passed; backend suite passed 283 tests; strict requirement and generation evaluations passed at 100; OpenAPI export and generated frontend API contract check passed; execution-runtime Playwright list check passed with 0 tests. |
