+++ frontend/src/components/projects/ProjectWorkspacePanel.jsx
@@ -1,121 +1,4 @@
 import OrchestratorCockpitPanel from "./OrchestratorCockpitPanel";
-import RunDetailsDrawer from "./RunDetailsDrawer";
-
-const STAGES = [
-	{ key: "requirements", label: "Requirements" },
-	{ key: "context", label: "Context" },
-	{ key: "use_cases", label: "Use Cases" },
-	{ key: "impact_analysis", label: "Impact Analysis" },
-	{ key: "test_cases", label: "Test Cases" },
-	{ key: "execution", label: "Execution" },
-	{ key: "reports", label: "Reports" },
-];
-
-const formatDateTime = (value) => {
-	if (!value) return "";
-	const date = new Date(value);
-	if (Number.isNaN(date.getTime())) return "";
-	return date.toLocaleString(undefined, {
-		month: "short",
-		day: "numeric",
-		hour: "2-digit",
-		minute: "2-digit",
-	});
-};
-
-function StagePill({ stage, state }) {
-	const version = state?.version || 0;
-	const className = state?.stale ? "project-stage-pill stale" : version ? "project-stage-pill ready" : "project-stage-pill";
-	return (
-		<div className={className}>
-			<span className="project-stage-label">{stage.label}</span>
-			<span className="project-stage-meta">
-				{version ? `v${version}` : "Not started"}
-				{state?.stale ? " · stale" : state?.approved ? " · approved" : ""}
-			</span>
-		</div>
-	);
-}
-
-function ExecutionHistory({ runs }) {
-	if (!runs?.length) {
-		return null;
-	}
-	return (
-		<div className="project-history-block">
-			<h3>Execution Runs</h3>
-			<div className="project-run-list">
-				{runs.slice(0, 4).map((run) => (
-					<div className="project-run-row" key={run.run_record_id}>
-						<span>{run.target_environment}</span>
-						<strong>{run.status}</strong>
-						<span>
-							{run.summary?.passed || 0} passed / {run.summary?.failed || 0} failed
-						</span>
-					</div>
-				))}
-			</div>
-		</div>
-	);
-}
-
-function ReportEvidence({ reportState, reportSnapshot }) {
-	if (!reportSnapshot) {
-		return null;
-	}
-	const payload = reportSnapshot.payload || {};
-	const evidence = payload.evidence || {};
-	const sourceSnapshotIds = evidence.source_snapshot_ids || {};
-	const executionRunIds = evidence.execution_run_ids || [];
-	const sourceEntries = Object.entries(sourceSnapshotIds).filter(([, value]) => value);
-	const status = reportState?.stale ? "Stale" : reportState?.approved ? "Approved" : "Draft";
-
-	return (
-		<div className="project-history-block">
-			<h3>Latest Report</h3>
-			<div className="project-run-list">
-				<div className="project-run-row">
-					<span>{payload.format || reportState?.operation || "report"}</span>
-					<strong>{status}</strong>
-					<span>{reportSnapshot.snapshot_id}</span>
-				</div>
-				{sourceEntries.slice(0, 3).map(([stage, snapshotId]) => (
-					<div className="project-run-row" key={`${stage}-${snapshotId}`}>
-						<span>{stage.replaceAll("_", " ")}</span>
-						<strong>Evidence</strong>
-						<span>{snapshotId}</span>
-					</div>
-				))}
-				{executionRunIds.slice(0, 2).map((runId) => (
-					<div className="project-run-row" key={runId}>
-						<span>execution run</span>
-						<strong>Evidence</strong>
-						<span>{runId}</span>
-					</div>
-				))}
-			</div>
-		</div>
-	);
-}
-
-function TimelinePreview({ events }) {
-	if (!events?.length) {
-		return null;
-	}
-	return (
-		<div className="project-history-block">
-			<h3>Timeline</h3>
-			<div className="project-timeline-list">
-				{events.slice(0, 4).map((event) => (
-					<div className="project-timeline-row" key={event.event_id}>
-						<span>{formatDateTime(event.occurred_at)}</span>
-						<strong>{event.summary}</strong>
-					</div>
-				))}
-			</div>
-		</div>
-	);
-}
 
 export default function ProjectWorkspacePanel({
 	currentProject,
@@ -124,19 +7,14 @@ export default function ProjectWorkspacePanel({
 	isLoadingProjects,
 	isCreatingProject,
 	orchestratorStatus,
-	orchestratorRuns,
 	isLoadingOrchestrator,
 	orchestratorError,
 	authActionDisabled,
 	onCreateProject,
 	onRefreshProjects,
-	onRefreshOrchestrator,
 	onOrchestratorAction,
 	orchestratorActionBusy,
 }) {
-	const stageState = currentProject?.stage_state || {};
-	const reportSnapshot = currentProject?.current_snapshots?.reports || null;
-	const hasProjectHistory = Boolean(reportSnapshot || currentProject?.execution_runs?.length || currentProject?.timeline?.length);
 	return (
 		<section className="project-workspace">
 			<div className="project-workspace-header">
@@ -164,39 +42,15 @@ export default function ProjectWorkspacePanel({
 				</button>
 			</div>
 
-			{currentProject && (
-				<>
-					<details className="project-progress-disclosure">
-						<summary>
-							<span>Project progress</span>
-							<span>{STAGES.filter((stage) => stageState[stage.key]?.version).length} stages started</span>
-						</summary>
-						<div className="project-stage-grid">
-							{STAGES.map((stage) => (
-								<StagePill key={stage.key} stage={stage} state={stageState[stage.key]} />
-							))}
-						</div>
-					</details>
-					<OrchestratorCockpitPanel
-						currentProject={currentProject}
-						status={orchestratorStatus}
-						runsPayload={orchestratorRuns}
-						isLoading={isLoadingOrchestrator}
-						error={orchestratorError}
-						authActionDisabled={authActionDisabled}
-						actionBusy={orchestratorActionBusy}
-						onRefresh={onRefreshOrchestrator}
-						onAction={onOrchestratorAction}
-					/>
-					<RunDetailsDrawer title="Project evidence" summary="Latest report, execution runs, and timeline" defaultOpen={hasProjectHistory}>
-						<div className="project-history-grid">
-							<ReportEvidence reportState={stageState.reports} reportSnapshot={reportSnapshot} />
-							<ExecutionHistory runs={currentProject.execution_runs || []} />
-							<TimelinePreview events={currentProject.timeline || []} />
-						</div>
-					</RunDetailsDrawer>
-				</>
-			)}
+			<OrchestratorCockpitPanel
+				currentProject={currentProject}
+				status={orchestratorStatus}
+				isLoading={isLoadingOrchestrator}
+				error={orchestratorError}
+				authActionDisabled={authActionDisabled}
+				actionBusy={orchestratorActionBusy}
+				onAction={onOrchestratorAction}
+			/>
 		</section>
 	);
 }
